.docker_requirements_scientific.txt
click
mypy
notebook
numpy
pandas
pre-commit
python-dotenv
pyyaml
scikit-learn
statsmodels
torch
transformers
